Refactoring

+++ src/native.c
@@ -16,4 +16,4 @@ NativeDefinition native_defs[] = {
     // clang-format on
 };
 
-int native_defs_length() { return sizeof(native_defs) / sizeof(NativeDefinition); }
+size_t native_defs_length() { return sizeof(native_defs) / sizeof(NativeDefinition); }
